+++ R/static.R
@@ -31,22 +31,26 @@ create_server = function(..., handler, ws_open = function(ws) NULL) {
 #' run this function, we can go to \samp{http://localhost:port} to browse the
 #' web pages either created from R or read from HTML files.
 #'
-#' \code{httd()} is a pure static server, and \code{httw()} is similar but
-#' watches for changes under the directory: if an HTML file is being viewed in
-#' the browser, and any files are modified under the directory, the HTML page
-#' will be automatically refreshed.
+#' \code{httd()} is a static file server by default (its \code{response}
+#' argument can turn it into a dynamic file server), and \code{httw()} is
+#' similar but watches for changes under the directory: if an HTML file is being
+#' viewed in the browser, and any files are modified under the directory, the
+#' HTML page will be automatically refreshed.
 #' @inheritParams server_config
 #' @param ... server configurations passed to \code{\link{server_config}()}
+#' @param response A function of the form \code{function(path, res, ...)} that
+#'   takes a file path and server response as input, and return a new response.
+#'   This can be useful for post-processing the response (for experts only).
 #' @export
 #' @references \url{https://github.com/yihui/servr}
 #' @examplesIf interactive()
 #' servr::httd()
-httd = function(dir = '.', ...) {
+httd = function(dir = '.', ..., response = NULL) {
   dir = normalizePath(dir, mustWork = TRUE)
   if (dir != '.') {
     owd = setwd(dir); on.exit(setwd(owd))
   }
-  create_server(dir, ..., handler = serve_dir(dir))
+  create_server(dir, ..., handler = serve_dir(dir, response))
 }
 
 #' @param watch a directory under which \code{httw()} is to watch for changes;
@@ -171,6 +175,7 @@ server_config = function(
   }
   port = as.integer(port)
   if (missing(daemon)) daemon = getOption('servr.daemon', ('-d' %in% cargs) || interactive())
+  if (!is.numeric(interval)) interval = as.numeric(interval)
   # rstudio viewer cannot display a page served at 0.0.0.0; use 127.0.0.1 instead
   host2 = if (host == '0.0.0.0' && is_rstudio()) '127.0.0.1' else host
   url = sprintf('http://%s:%d', hosturl(host2), port)
@@ -228,7 +233,7 @@ modify_path = function(req, baseurl) {
   req
 }
 
-serve_dir = function(dir = '.') function(req) {
+serve_dir = function(dir = '.', response = NULL) function(req) {
   owd = setwd(dir); on.exit(setwd(owd), add = TRUE)
   path = decode_path(req)
   status = 200L
@@ -276,7 +281,7 @@ serve_dir = function(dir = '.') function(req) {
       b2 = as.numeric(range[2])
       if (length(range) == 2 && range[1] == "bytes") {
         # open-ended range request
-        # e.g. Chrome sends the range reuest 'bytes=0-'
+        # e.g. Chrome sends the range request 'bytes=0-'
         # http://stackoverflow.com/a/18745164/559676
         range[3] = file_size(path) - 1
       }
@@ -299,11 +304,12 @@ serve_dir = function(dir = '.') function(req) {
     }
   }
   if (is.character(body) && length(body) > 1) body = paste2(body)
-  list(
+  res = list(
     status = status, body = body,
     headers = c(list('Content-Type' = type), if (status == 206L) list(
       'Content-Range' = paste0("bytes ", range[2], "-", range[3], "/", file_size(path))
       ),
       'Accept-Ranges' = 'bytes') # indicates that the server supports range requests
   )
+  if (is.function(response)) response(path, res) else res
 }